requirements.txt
torch>=2.1.0
transformers>=4.38.0
sglang
pyyaml>=6.0
tqdm>=4.66.0
requests>=2.31.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
pydantic
orjson
uvicorn
uvloop
fastapi
psutil
zmq
sentencepiece
compressed_tensors
einops
dill
msgspec
partial_json_parser
python-multipart
torchao
xgrammar
